+++ xgboost/src/main/java/hivemall/xgboost/XGBoostBaseUDTF.java
@@ -25,7 +25,9 @@ import hivemall.utils.hadoop.HiveUtils;
 import hivemall.utils.math.MathUtils;
 import hivemall.xgboost.utils.DMatrixBuilder;
 import hivemall.xgboost.utils.DenseDMatrixBuilder;
+import hivemall.xgboost.utils.NativeLibLoader;
 import hivemall.xgboost.utils.SparseDMatrixBuilder;
+import hivemall.xgboost.utils.XGBoostUtils;
 import matrix4j.utils.lang.ArrayUtils;
 import matrix4j.utils.lang.Primitives;
 import ml.dmlc.xgboost4j.java.Booster;
