+++ xgboost/src/main/java/hivemall/xgboost/utils/NativeLibLoader.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package hivemall.xgboost;
+package hivemall.xgboost.utils;
 
 import java.io.File;
 import java.io.FileInputStream;
