+++ xgboost/src/main/java/hivemall/xgboost/utils/XGBoostUtils.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package hivemall.xgboost;
+package hivemall.xgboost.utils;
 
 import ml.dmlc.xgboost4j.LabeledPoint;
 import ml.dmlc.xgboost4j.java.Booster;
